Significantly improved the Rules and remove unnecessary redundant rules
Completely changed the structure of the rules added
1) General Rules:
core.mdc and absolute-no-rights.mdc as Always Apply Rules

refresh.mdc, request.mdc and retro.mdc as optional to apply manually.

you can find a brief about those roles in:

https://gist.github.com/aashari/07cc9c1b6c0debbeb4f4d94a3a81339e

2) Reduced the amount of information given to the agent through rules and only sticked to 3 rule files which are mcp-tools-reference.mdc, step-nodes-reference.mdc and unity-mcp-essentials.mdc

Observation: the performance of the model has improved significantly due to the rules we added make sure to read the link to know how to use the rules to optimize the performance of the model.